+++ src/compas_viewer/ui/toolbar.py
@@ -1,5 +1,6 @@
 from compas_viewer.base import Base
 from compas_viewer.components.button import Button
+from compas_viewer.components.combo_box import ViewComboBox
 
 
 def test_action() -> None:
@@ -15,4 +16,5 @@ class ToolBar(Base):
         self.widget.setMovable(False)
         self.widget.setObjectName("Tools")
         self.widget.setHidden(not self.viewer.config.ui.toolbar.show)
+        self.widget.addWidget(ViewComboBox())
         self.widget.addWidget(Button("zoom_selected.svg", "zoom", test_action()))