+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `compas_viewer.components.combo_box` component
+* Added `ViewComboBox` component handle viewmode change
+
 ### Changed
 
 ### Removed